Move extension to Player module
