+++ pillarbox-player/src/main/java/ch/srgssr/pillarbox/player/Extensions.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2023. SRG SSR. All rights reserved.
  * License information is available from the LICENSE file.
  */
-package ch.srgssr.pillarbox.ui
+package ch.srgssr.pillarbox.player
 
 import androidx.media3.common.VideoSize
